restboty-form/miapp/requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
backports.zoneinfo==0.2.1
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
dataclasses==0.6
dill==0.3.8
Django==4.2.15
djangorestframework==3.15.2
djongo==1.2.31
dnspython==2.6.1
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.4.0
grpcio==1.65.5
h5py==3.11.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
Jinja2==3.1.4
joblib==1.4.2
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
mongoengine==0.28.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.1
protobuf==4.25.4
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.4
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sqlparse==0.5.1
sympy==1.13.2
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.5.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.5.0
yarl==1.9.4
zipp==3.20.0
